training/requirements.txt
# requirements.txt

asttokens==2.4.1
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
coverage==7.6.1
cycler==0.12.1
dataclasses-json==0.5.14
debugpy==1.8.5
decorator==5.1.1
executing==2.1.0
filelock==3.13.1
filetype==1.2.0
fonttools==4.53.1
fsspec==2024.2.0
idna==3.7
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.27.0
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.7
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
mpmath==1.3.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.23.5
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==1.4.4
parso==0.8.4
Pillow==9.3.0
platformdirs==4.3.3
pluggy==1.5.0
prompt_toolkit==3.0.47
psutil==6.0.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pycocotools==2.0.8
Pygments==2.18.0
pyparsing==3.1.4
pysolotools==0.3.17
pytest==8.3.3
pytest-cov==5.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyWavelets==1.7.0
pywin32==306
PyYAML==6.0.2
pyzmq==26.2.0
ratelimit==2.2.1
requests==2.32.3
requests-toolbelt==1.0.0
responses==0.25.3
roboflow==1.1.44
scipy==1.9.3
seaborn==0.13.2
six==1.16.0
stack-data==0.6.3
sympy==1.12
thop==0.1.1.post2209072238
# torch==2.4.1+cu124 # get from https://pytorch.org/
# torchaudio==2.4.1+cu124
# torchvision==0.19.1+cu124
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.97
ultralytics-thop==2.0.6
urllib3==2.2.3
wcwidth==0.2.13
